+++ ios/Classes/SwiftGoSellSdkFlutterPlugin.swift
@@ -1,7 +1,7 @@
 import Flutter
 import UIKit
 import goSellSDK
-import TapCardValidator
+import TapCardVlidatorKit_iOS
 
 public class SwiftGoSellSdkFlutterPlugin: NSObject, FlutterPlugin {
     let session = Session()
@@ -343,21 +343,42 @@ extension SwiftGoSellSdkFlutterPlugin: SessionDataSource {
       if let cardTypeString:String = argsSessionParameters?["allowedCadTypes"] as? String {
         let cardTypeComponents: [String] = cardTypeString.components(separatedBy: ".")
         if cardTypeComponents.count == 2 {
-          var cardType:cardTypes = .All
-          cardTypes.allCases.forEach{
-            if $0.description.lowercased() == cardTypeComponents[1].lowercased() {
-              cardType = $0
+            let cardType = CardType(cardTypeString: cardTypeComponents[1])
+            if !cardType.isEqual(CardType.init(cardType:.All)) {
+                return [cardType]
             }
-          }
-          if cardType == .All {
-            return [CardType(cardType: .Debit), CardType(cardType: .Credit)]
-          }else
-          {
-            return [CardType(cardType: cardType)]
-          }
         }
       }
       return [CardType(cardType: .Debit), CardType(cardType: .Credit)]
+    }
+    
+    public var paymentMetadata: Metadata? {
+        
+        if let paymentMetaDataString = argsSessionParameters?["paymentMetaData"] as? String, let data = paymentMetaDataString.data(using: .utf8) {
+            
+                
+                do {
+                    let decoder = JSONDecoder()
+                    let paymentMetaDataValue:Metadata = try decoder.decode(Metadata.self, from: data)
+                    return paymentMetaDataValue
+                  } catch {
+                    print(error.localizedDescription)
+                    return nil
+                  }
+            
+        }
+        
+        
+        if let paymentMetaDataString = argsSessionParameters?["paymentMetaData"] as? [String: Any] {
+            
+            let converted = paymentMetaDataString.compactMapValues { $0 as? String }
+            return converted
+              
+        }
+        
+        return nil
+       
+
     }
 }